Compute interaction effect sizes

TODO: !

Usage

inter_effsize(
  effsize,
  colnames,
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean,
  B_sd,
  B_n,
  AB_mean,
  AB_sd,
  AB_n,
  data
)

Arguments

effsize

Character. Effect‐size metric to compute. Supported: "lnrr".

colnames

Character vector of length 2 giving the names to use for the factors A and B

Ctrl_mean, Ctrl_sd, Ctrl_n,

A_mean,A_sd,A_n, B_mean,B_sd,B_n, AB_mean,AB_sd,AB_n Unquoted column names identifying group means, SDs, and sample sizes for the control (Ctrl), single-factor treatments A and B, and the combined treatment AB.

data

A data frame containing all columns listed above.

Value

The original data with the effect size and its variance appended, named as in colnames.

Author

Facundo Decunta - fdecunta@agro.uba.ar

Examples

if (FALSE) { # \dontrun{
# This example won't be run during R CMD check
result <- inter_effsize(
  effsize   = "lnrr",
  colnames  = c("yi", "vi"),
  Ctrl_mean = ctrl_mean, Ctrl_sd = ctrl_sd, Ctrl_n = ctrl_n,
  A_mean    = a_mean,    A_sd    = a_sd,    A_n    = a_n,
  B_mean    = b_mean,    B_sd    = b_sd,    B_n    = b_n,
  AB_mean   = ab_mean,   AB_sd   = ab_sd,   AB_n   = ab_n,
  data      = df
)
} # }