Interaction effect: Standardized mean difference

Computes the interaction effect between factors A and B in factorial data.

Usage

.interaction_SMD(
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean,
  B_sd,
  B_n,
  AB_mean,
  AB_sd,
  AB_n,
  hedges_correction = TRUE
)

Arguments

Ctrl_mean

Mean outcome from the Control treatment

Ctrl_sd

Standard deviation from the control treatment

Ctrl_n

Sample size from the control streatment

A_mean

Mean outcome from the treatment

A_sd

Standard deviation from the treatment

A_n

Sample size from the treatment

B_mean

Mean outcome from the B treatment

B_sd

Standard deviation from the B treatment

B_n

Sample size from the B treatment

AB_mean

Mean outcome from the interaction AxB treatment

AB_sd

Standard deviation from the interaction AxB treatment

AB_n

Sample size from the interaction AxB treatment

hedges_correction

Logical. Apply or not Hedges' correction for small-sample bias. Default is TRUE

Details

See the package vignette for a detailed description of the formula.

References

Gurevitch, J., Morrison, J. A., & Hedges, L. V. (2000). The interaction between competition and predation: a meta-analysis of field experiments. The American Naturalist, 155(4), 435-453.

Morris, W. F., Hufbauer, R. A., Agrawal, A. A., Bever, J. D., Borowicz, V. A., Gilbert, G. S., ... & Vázquez, D. P. (2007). Direct and interactive effects of enemies and mutualists on plant performance: a meta‐analysis. Ecology, 88(4), 1021-1029. https://doi.org/10.1890/06-0442